# src/app/backend/resources/swagger/cryptocurrency-open-api.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.3",
    "title": "CNK Cryptocurrency",
    "description": "API to manage CNK cryptocurrency",
    "contact": {
      "name": "Juan Sebastián Bravo",
      "email": "jsbravoc@gmail.com",
      "url": "https://jsbravoc.com"
    },
    "license": {
      "name": "BSD-3-Clause",
      "url": "https://opensource.org/licenses/BSD-3-Clause"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "paths": {
    "/users": {
      "get": {
        "tags": ["users"],
        "summary": "Retrieves all users and their transactions registered in the blockchain.",
        "description": "Retrieves all users registered in the blockchain. Note that, by default, Hyperledger Sawtooth REST API will return up to 100 users.",
        "operationId": "getUsers",
        "parameters": [
          {
            "name": "expanded",
            "in": "query",
            "description": "Expand recursively the transactions",
            "required": false,
            "schema": {
              "type": "boolean",
              "format": "boolean",
              "example": true
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit queried users",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "example": 100
            }
          },
          {
            "name": "hidePublicKey",
            "in": "query",
            "description": "Hide users' public key.",
            "required": false,
            "schema": {
              "type": "boolean",
              "format": "boolean",
              "example": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieves all the users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/ArrayOfExtendedUsers"
                }
              }
            }
          },
          "503": {
            "description": "Sawtooth service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Sawtooth503Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["users"],
        "summary": "Creates a new user",
        "description": "Allows the creation of a new user inside the blockchain. Note that cryptocurrency-sensitive properties such as `balance`, `latest_transactions` and `pending_transactions` cannot be set upon creation, thus they will default to 0 and empty arrays respectively.",
        "operationId": "postUser",
        "requestBody": {
          "description": "User to create in the system",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/CreateUserInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created user successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/CreatedUser"
                }
              }
            },
            "links": {
              "GetUserByAddress": {
                "operationId": "getUserByAddress",
                "parameters": {
                  "address": "$response.body#/payload/address"
                },
                "description": "The `address` value returned in the response's payload can be used as the `address` parameter in `GET /users/{address}`"
              }
            }
          },
          "400": {
            "description": "Input validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Logic111bError"
                }
              }
            }
          },
          "503": {
            "description": "Sawtooth service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Sawtooth503Error"
                }
              }
            }
          }
        }
      }
    },
    "/users/{address}": {
      "get": {
        "tags": ["users"],
        "summary": "Retrieves one specific user.",
        "description": "Retrieves one specific user by their address. The user must exist in the blockchain, otherwise the endpoint's response will be a 400 error.",
        "operationId": "getUserByAddress",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "description": "The address of the user to retrieve",
            "required": true,
            "schema": {
              "type": "string",
              "format": "string",
              "example": "jsbravoc"
            }
          },
          {
            "name": "hidePublicKey",
            "in": "query",
            "description": "Hide user's public key.",
            "required": false,
            "schema": {
              "type": "boolean",
              "format": "boolean",
              "example": true
            }
          },
          {
            "name": "expanded",
            "in": "query",
            "description": "Expand recursively the user's transactions",
            "required": false,
            "schema": {
              "type": "boolean",
              "format": "boolean",
              "example": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieves the specified user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/ExtendedUserNoTransactions"
                }
              }
            }
          },
          "400": {
            "description": "Input validation error: User does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Logic111bErrorAddress"
                }
              }
            }
          },
          "503": {
            "description": "Sawtooth service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Sawtooth503Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["users"],
        "summary": "Updates one specific user.",
        "description": "Allows the update of specific user properties: `role`, `description`, `return_to` addresses, `permissions` and `active`. Note that cryptocurrency-sensitive properties such as `address`, `public_key`, `balance`, `latest_transactions` and `pending_transactions` cannot be changed/updated.",
        "operationId": "updateUser",
        "requestBody": {
          "description": "Properties to update",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/UpdateUserInput"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "description": "The address of the user to update",
            "required": true,
            "schema": {
              "type": "string",
              "format": "string",
              "address": "jsbravoc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updates user information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/UpdatedUser"
                }
              }
            }
          },
          "400": {
            "description": "Input validation error: User does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Logic111bError"
                }
              }
            }
          },
          "503": {
            "description": "Sawtooth service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Sawtooth503Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["users"],
        "summary": "Sets a user as inactive and moves all their cryptocurrency balance to another user specified in this `return_to` addresses.",
        "description": "Sets as inactive one specific user and transfers their CNKs to another user, specified in their `return_to` addresses. Since `return_to` addresses is a dictionary composed by {%reason%: %address%}, the user must have the specified %reason% in the query parameters defined in their `return_to` addresses to be able to use the endpoint with that %reason%.",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "description": "The address of the user to update",
            "required": true,
            "schema": {
              "type": "string",
              "format": "public_key"
            }
          },
          {
            "name": "reason",
            "in": "path",
            "description": "The reason why the user is being deleted. Must match any return_to reason of the user.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The user was deleted and the user's CNKs were transferred successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/CreatedTransaction"
                }
              }
            }
          },
          "400": {
            "description": "Input validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Logic111bError"
                }
              }
            }
          },
          "503": {
            "description": "Sawtooth service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Sawtooth503Error"
                }
              }
            }
          }
        }
      }
    },
    "/cryptocurrency": {
      "get": {
        "tags": ["cryptocurrency"],
        "summary": "Retrieves all transactions registered in the blockchain.",
        "description": "Retrieves all transactions registered in the blockchain. Note that, by default, Hyperledger Sawtooth REST API will return up to 100 transactions.",
        "operationId": "getCryptocurrency",
        "parameters": [
          {
            "name": "expanded",
            "in": "query",
            "description": "Expand recursively the transactions",
            "required": false,
            "schema": {
              "type": "boolean",
              "format": "boolean",
              "example": true
            }
          },
          {
            "name": "hidePending",
            "in": "query",
            "description": "Hide pending transactions",
            "required": false,
            "schema": {
              "type": "boolean",
              "format": "boolean",
              "example": true
            }
          },
          {
            "name": "hideInvalid",
            "in": "query",
            "description": "Hide invalid transactions",
            "required": false,
            "schema": {
              "type": "boolean",
              "format": "boolean",
              "example": true
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit queried transactions",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "example": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Information about the cryptocurrency flow.",
            "content": {
              "application/json": {
                "examples": {
                  "Transaction array": {
                    "$ref": "#/examples/NonExpandedTransactionArray"
                  },
                  "Expanded transaction array": {
                    "$ref": "#/examples/ExpandedTransactionArray"
                  }
                }
              }
            }
          },
          "503": {
            "description": "Sawtooth service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Sawtooth503Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["cryptocurrency"],
        "summary": "Creates a new transaction",
        "description": "Allows the creation of a new transaction, which can be a coinbase (no-input) transaction or a normal transaction. A coinbase transaction must not have any sender defined and also its recipient must have **coinbase** permission set to _true_. This endpoint validates balance and also existing & active users, thus `sender` must have the specified balance and both the `sender` and the `recipient` must exist and be active. **Note:** When ENV variable `DISABLE_INTEGRITY_VALIDATION` is set to _false_, the signature of the transaction must be created with the user's private key (thus, decryptable with their public_key) and its contents must match the transactions sensitive info (`sender`, `recipient`, `amount`, `creator`).",
        "operationId": "createTransaction",
        "requestBody": {
          "description": "New transaction to add to the blockchain",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/SimpleTransaction"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transaction created successfully",
            "content": {
              "application/json": {
                "examples": {
                  "Transaction with supporting_transactions": {
                    "$ref": "#/examples/CreatedTransaction"
                  },
                  "Coinbase transaction": {
                    "$ref": "#/examples/CreatedCoinbaseTransaction"
                  }
                }
              }
            },
            "links": {
              "GetTransactionByAddress": {
                "operationId": "getTransactionByAddress",
                "parameters": {
                  "address": "$response.body#/payload/signature"
                },
                "description": "The `signature` value returned in the response's payload can be used as the `address` parameter in `GET /cryptocurrency/{address}`"
              }
            }
          },
          "400": {
            "description": "Input validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Logic111bErrorRecipient"
                }
              }
            }
          },
          "503": {
            "description": "Sawtooth service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/Sawtooth503Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "SimpleTransaction": {
      "type": "object",
      "required": ["amount", "sender"],
      "properties": {
        "sender": {
          "type": "number",
          "format": "double",
          "example": "jsbravoc"
        },
        "amount": {
          "type": "number",
          "format": "double",
          "example": "2.8"
        },
        "recipient": {
          "type": "string",
          "format": "string",
          "example": "jsbravoc2"
        },
        "pending": {
          "type": "boolean",
          "format": "boolean",
          "example": false
        },
        "signature": {
          "type": "string",
          "format": "string",
          "example": "0xc79eec8f62d09794feeb0677f2330d70cee5c640a03b2f35e508f8fe12140ff8754d37904ed443d942363ad3ebbd5f006536794f4110355343c0b91f4f2a48221b"
        },
        "valid_thru": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
    "UpdateUserInput": {
      "type": "object",
      "properties": {
        "role": {
          "type": "string",
          "format": "string",
          "example": "Student"
        },
        "description": {
          "type": "string",
          "format": "string",
          "example": "Developer of this API"
        },
        "return_to": {
          "$ref": "#/definitions/ReturnTo"
        },
        "permissions": {
          "$ref": "#/definitions/Permissions"
        }
      }
    },
...
    "111bError": {
      "type": "object",
      "required": ["errorCode", "msg", "error"],
      "properties": {
        "errorCode": {
          "type": "string",
          "format": "string",
          "example": "111b"
        },
        "msg": {
          "type": "string",
          "format": "string",
          "example": "User with address created.user does not exist"
        },
        "error": {
          "type": "string",
          "format": "string",
          "example": "The user created.user does not exist"
        },
        "location": {
          "type": "string",
          "format": "string",
          "example": "body"
        },
        "param": {
          "type": "string",
          "format": "string",
          "example": "return_to"
        },
        "value": {
          "type": "string",
          "format": "string",
          "example": "created.user"
        }
      }
    },
    "111bErrorAddress": {
      "type": "object",
      "required": ["errorCode", "msg", "error"],
      "properties": {
        "errorCode": {
          "type": "string",
          "format": "string",
          "example": "111b"
        },
        "msg": {
          "type": "string",
          "format": "string",
          "example": "User with address created.user does not exist"
        },
        "error": {
          "type": "string",
          "format": "string",
          "example": "The user created.user does not exist"
        },
        "location": {
          "type": "string",
          "format": "string",
          "example": "query"
        },
        "param": {
          "type": "string",
          "format": "string",
          "example": "address"
        },
        "value": {
          "type": "string",
          "format": "string",
          "example": "created.user"
        }
      }
    },
...
    "503Error": {
      "type": "object",
      "required": ["errorCode", "msg", "error"],
      "properties": {
        "errorCode": {
          "type": "string",
          "format": "string",
          "example": "503"
        },
        "msg": {
          "type": "string",
          "format": "string",
          "example": "Hyperledger Sawtooth REST API is temporarily unavailable"
        },
        "error": {
          "type": "string",
          "format": "string",
          "example": "Hyperledger Sawtooth REST API is temporarily unavailable"
        }
      }
    },
...
    "CreatedUser": {
      "type": "object",
      "properties": {
        "msg": {
          "type": "string",
          "format": "string",
          "example": "User created"
        },
        "payload": {
          "$ref": "#/definitions/ExtendedUserNoTransactions"
        }
      }
    },
    "CreatedTransaction": {
      "type": "object",
      "properties": {
        "msg": {
          "type": "string",
          "format": "string",
          "example": "Transaction created"
        },
        "payload": {
          "$ref": "#/definitions/NonExpandedTransaction"
        }
      }
    },
    "UpdatedUser": {
      "type": "object",
      "properties": {
        "msg": {
          "type": "string",
          "format": "string",
          "example": "User updated"
        },
        "payload": {
          "$ref": "#/definitions/ExtendedUser"
        }
      }
    },
...
    "ExtendedUser": {
      "type": "object",
      "properties": {
        "active": {
          "type": "boolean",
          "format": "boolean",
          "example": "true"
        },
        "balance": {
          "type": "number",
          "format": "double",
          "example": 0
        },
        "address": {
          "type": "string",
          "format": "string",
          "example": "jsbravoc"
        },
        "description": {
          "type": "string",
          "format": "string",
          "example": "Developer of this API"
        },
        "return_to": {
          "type": "object"
        },
        "latest_transactions": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "string",
            "example": "0xc79eec8f62d09794feeb0677f2330d70cee5c640a03b2f35e508f8fe12140ff8754d37904ed443d942363ad3ebbd5f006536794f4110355343c0b91f4f2a48221b"
          }
        },
        "pending_transactions": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "string",
            "example": "0xcsa8f62d09sasfb0677f2330d70cee5c640a03b2f35e508f8fe12140ff8754d37904ed443d942363ad3ebbd5f006536794f4110355343c0b91f4f2a48221b"
          }
        },
        "public_key": {
          "type": "string",
          "format": "string",
          "example": "023758f7c69eef837bb3ed6d684059b74272538df147e3a56705f80b5649dae6d2"
        },
        "permissions": {
          "$ref": "#/definitions/Permissions"
        }
      }
    },
    "ExtendedUserNoTransactions": {
      "type": "object",
      "properties": {
        "active": {
          "type": "boolean",
          "format": "boolean",
          "example": "true"
        },
        "balance": {
          "type": "number",
          "format": "double",
          "example": 0
        },
        "address": {
          "type": "string",
          "format": "string",
          "example": "jsbravoc"
        },
        "description": {
          "type": "string",
          "format": "string",
          "example": "Developer of this API"
        },
        "return_to": {
          "type": "object"
        },
        "latest_transactions": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "string",
            "example": "0xc79eec8f62d09794feeb0677f2330d70cee5c640a03b2f35e508f8fe12140ff8754d37904ed443d942363ad3ebbd5f006536794f4110355343c0b91f4f2a48221b"
          },
          "example": []
        },
        "pending_transactions": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "string",
            "example": "0xc79eec8fdass4feeb0677f2330d70cee5c640a03b2f35e508f8fe12140ff8754d37904ed443d942363ad3ebbd5f006536794f4110355343c0b91f4f2a48221b"
          },
          "example": []
        },
        "public_key": {
          "type": "string",
          "format": "string",
          "example": "023758f7c69eef837bb3ed6d684059b74272538df147e3a56705f80b5649dae6d2"
        },
        "permissions": {
          "$ref": "#/definitions/Permissions"
        }
      }
    },
    "CreateUserInput": {
      "type": "object",
      "properties": {
        "return_to": {
          "$ref": "#/definitions/ReturnTo"
        },
        "address": {
          "type": "string",
          "format": "string",
          "example": "jsbravoc"
        },
        "public_key": {
          "type": "string",
          "format": "string",
          "example": "023758f7c69eef837bb3ed6d684059b74272538df147e3a56705f80b5649dae6d2"
        },
        "role": {
          "type": "string",
          "format": "string",
          "example": "Student"
        },
        "description": {
          "type": "string",
          "format": "string",
          "example": "Developer of this API"
        },
        "permissions": {
          "$ref": "#/definitions/Permissions"
        }
      }
    },
    "Permissions": {
      "type": "object",
      "properties": {
        "coinbase": {
          "type": "boolean",
          "format": "boolean",
          "example": "true"
        },
        "transfer_to": {
          "$ref": "#/definitions/TransferTo"
        }
      }
    },
    "TransferTo": {
      "type": "object",
      "properties": {
        "created.user": {
          "type": "boolean",
          "format": "boolean",
          "example": "false"
        }
      }
    },
    "ReturnTo": {
      "type": "object",
      "properties": {
        "reason_to_return": {
          "type": "string",
          "format": "string",
          "example": "created.user"
        }
      }
    },
    "Sawtooth503Error": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/503Error"
          }
        }
      }
    },
    "Logic111bError": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/111bError"
          }
        }
      }
    },
    "Logic111bErrorAddress": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/111bErrorAddress"
          }
        }
      }
    },
    "Logic111bErrorRecipient": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/111bErrorRecipient"
          }
        }
      }
    },
    "ArrayOfExtendedUsers": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/ExtendedUser"
      }
    },
...
  },
...
}